+++ NEWS.md
@@ -1,3 +1,9 @@
+# zerenebatchR 0.2.0
+
+* Added option to create export directory if it doesn't exist.
+* Added `expand_zs_dataframe` function to expand data.frame specifying the first 
+and last photos in a sequence.
+
 # zerenebatchR 0.1.0
 
 * Added a `NEWS.md` file to track changes to the package.